Module nitride::controllers::email_controllers

email controllers : contains handlers (controllers) to send, receive emails, reply email e.t.

Functions

archive_email

move email to archive receive the email id as path variable find the email star it return the update or error

delete_email

delete email receive the id of the mail to delete exec the query on the database return result

get_all_emails

get all emails by pagination default to default pagination config

get_email_by_id

fetch email retrieve an email from the data store if the email was found, return the fond email else, return a not found error

receive_email

receive email being sent from the portfolio store it in the database

reply_email

reply email receive only the user email and subject and message send the message to the user

send_email

send email handler receive the user email, subject, fullname and message call on lettre to dispatch the mail to the user

star_email

star email receive the email id as path variable find the email star it return the update or error

un_star_email

unstRstar email receive the email id as path variable find the email star it return the update or error

Opaque Types

ğŸ”’

ğŸ”’

ğŸ”’

ğŸ”’

ğŸ”’

ğŸ”’

ğŸ”’

ğŸ”’

ğŸ”’