Module nitride::controllers::project_controllers

project controllers

Functions

add_project

create new project accept the following data

edit_project

edit project accept the project id as route parameter find the project effect edits return updated project object

get_all_projects

get all projects retrieve all project with pagination

get_project_by_id

get one project collect the project id from the client search the database for the project return success and response or 404 error

Opaque Types

🔒

🔒

🔒

🔒

🔒