Trait nitride::shared::api_response::EnumerateFields

pub trait EnumerateFields {
    fn collect_as_strings(&self) -> HashMap<String, String>;
}

a trait to return the field of the structs as an array of strings the implementation on user information will return the user is, firstname, username … on the user authentication struct, the implementation will return the user email and password

example

  //destructure the HTTP request body
  let UserInformation { fullname, password, username, email, } = &payload;

 // check through the fields to see that no field was badly formatted
 let entries = &payload.collect_as_strings();
 let mut bad_request_errors: Vec<String> = Vec::new();

 for (key, value) in entries {
  if value.is_empty() {
  let error = format!("{key} is empty");
  bad_request_errors.push(error);
 }
}

// do something with the response

Required Methods

fn collect_as_strings(&self) -> HashMap<String, String>